Refactor to a single job
